Minor restructuring of Project
