+++ validators.js
@@ -1,27 +1,6 @@
-/**
- * Return the true Datatype of a value
- * @param obj
- * @return {string}
- */
-let toType = (obj) => {
-    return ({}).toString.call(obj).match(/\s([a-zA-Z]+)/)[1].toLowerCase();
-};
+const {toType} = require('./utils');
 
-/**
- * Checks if 'obj' is a valid Schema, either a string defining the Datatype or and object containing the 'type' key
- * which is a string defining the Datatype
- * @param obj {string || object}
- * @return {boolean}
- */
-let isSchema = (obj) => {
-    if(toType(obj) === 'string')
-        return true;
-    else if(obj['type'] !== 'undefined')
-        return obj['type'] === 'string';
-    else
-        return false;
-};
-
+// noinspection JSUnusedGlobalSymbols
 /**
  * A list of validators. Can be expanded
  */
@@ -171,4 +150,4 @@ let applyValidators = (value, schema) => {
     return bool;
 };
 
-module.exports = {toType, isSchema, validators, applyValidators};
+module.exports = applyValidators;
